requirements.txt
appdirs==1.4.0
cycler==0.10.0
functools32==3.2.3.post2
nltk==3.2.2
numpy==1.12.0
nytimesarticle==0.1.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
six==1.10.0
subprocess32==3.2.7
wordcloud==1.2.1
